# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api backend_challeng_doc88_back",
        "version": "0.1"
    },
    "paths": {
...
        "/v1/client/all": {
            "get": {
                "summary": "Exibe todos os registros, exceto deletados",
                "operationId": "App\\Http\\Controllers\\ClientController::getAll",
                "responses": {
                    "200": {
                        "description": "Lista de clientes"
                    },
                    "500": {
                        "description": "Erro no formato ou tipo dos dados enviados"
                    },
                    "default": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/v1/client/{id}": {
            "get": {
                "summary": "Exibe um cliente ativo pelo {id}",
                "operationId": "App\\Http\\Controllers\\ClientController::getById",
                "parameters": [
                    {
                        "name": "show_deleted",
                        "in": "header",
                        "description": "{id} do cliente a ser localizado.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do cliente solicitado"
                    },
                    "500": {
                        "description": "Erro no formato ou tipo dos dados enviados"
                    },
                    "default": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "put": {
                "summary": "Atualiza os dados do cliente",
                "operationId": "App\\Http\\Controllers\\ClientController::update",
                "parameters": [
                    {
                        "name": "name",
                        "in": "header",
                        "description": "Valor para o campo 'name'.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Fulado"
                    },
                    {
                        "name": "email",
                        "in": "header",
                        "description": "Valor para o campo 'email'.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "teste@teste.com"
                    },
                    {
                        "name": "address",
                        "in": "header",
                        "description": "Valor para o campo 'address'.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Rua dos Tolos"
                    },
                    {
                        "name": "complement",
                        "in": "header",
                        "description": "Valor para o campo 'complement'.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Ap 25"
                    },
                    {
                        "name": "complement",
                        "in": "header",
                        "description": "Valor para o campo 'neighborhood' (bairro).",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Berrini"
                    },
                    {
                        "name": "phone",
                        "in": "header",
                        "description": "Valor para o campo 'phone'.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2345678"
                    },
                    {
                        "name": "cep",
                        "in": "header",
                        "description": "Valor para o campo 'cep'.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "00000001"
                    },
                    {
                        "name": "birth_date",
                        "in": "header",
                        "description": "Valor para o campo 'birth_date'.",
                        "required": false,
                        "schema": {
                            "type": "date"
                        },
                        "example": "1993-09-22"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do cliente atualizado."
                    },
                    "406": {
                        "description": "Erro no formato ou tipo dos dados enviados."
                    },
                    "default": {
                        "description": "Erro interno do servidor."
                    }
                }
            },
            "delete": {
                "summary": "Deleta um cliente pelo {id}",
                "operationId": "App\\Http\\Controllers\\ClientController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "header",
                        "description": "{id} do cliente a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensagem informando que o cliente foi deletado."
                    },
                    "500": {
                        "description": "Erro no formato ou tipo dos dados enviados."
                    },
                    "default": {
                        "description": "Erro interno do servidor."
                    }
                }
            }
        },
...
        "/v1/order/{id}": {
            "get": {
                "summary": "Exibe um pedido ativo pelo {id}",
                "operationId": "App\\Http\\Controllers\\OrderController::getById",
                "parameters": [
                    {
                        "name": "show_deleted",
                        "in": "header",
                        "description": "{id} do pedido a ser localizado.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do pedido solicitado"
                    },
                    "500": {
                        "description": "Erro no formato ou tipo dos dados enviados"
                    },
                    "default": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "put": {
                "summary": "Alterações no pedido não são autorizadas, você deve cancelá-lo e efetuar um novo pedido.",
                "operationId": "App\\Http\\Controllers\\OrderController::update",
                "responses": {
                    "404": {
                        "description": "Alterações no pedido não são autorizadas, você deve cancelá-lo e efetuar um novo pedido."
                    }
                }
            },
            "delete": {
                "summary": "Deleta um pedido pelo {id}",
                "operationId": "App\\Http\\Controllers\\OrderController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "header",
                        "description": "{id} do pedido a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensagem informando que o pedido foi deletado."
                    },
                    "500": {
                        "description": "Erro no formato ou tipo dos dados enviados."
                    },
                    "default": {
                        "description": "Erro interno do servidor."
                    }
                }
            }
        },
...
    }
}